+++ resources/js/admin/pages/quizzes/routes.tsx
@@ -4,6 +4,7 @@ import {getCourse} from "../../entities/course/api/queries/getCourse";
 import {BreadcrumbHandle} from "../../shared/ui/breadcrumbs/types";
 import {getQuiz} from "../../entities/quiz/queries/getQuiz";
 import {QuizEditPage} from "./edit/ui/QuizEditPage";
+import {QuizDetailsPage} from "./show/ui/QuizDetailsPage";
 
 export const quizRoutes: RouteObject[] = [
     {
@@ -50,6 +51,23 @@ export const quizRoutes: RouteObject[] = [
 
                 path: ":quizUuid/edit",
                 element: <QuizEditPage />,
+                loader: getQuiz
+            },
+
+            {
+                handle: {
+                    title: "Страница теста",
+                    breadcrumb: ({ data }) => {
+                        const quiz = data as Awaited<ReturnType<typeof getQuiz>>;
+
+                        return {
+                            label: quiz?.title,
+                        };
+                    }
+                } satisfies BreadcrumbHandle,
+
+                path: ":quizUuid",
+                element: <QuizDetailsPage />,
                 loader: getQuiz
             }
         ]