+++ resources/js/admin/shared/constants/links.ts
@@ -1,0 +1,1 @@
+export const noImageLink = '/assets/cms/img/no-image.jpg';